chore: Prevent deleting the last admin user

+++ packages/teg-auth/src/models/user/mod.rs
@@ -1,6 +1,7 @@
 use chrono::prelude::*;
 use juniper::{
     FieldResult,
+    FieldError,
     ID,
 };
 
@@ -54,6 +55,7 @@ impl User {
 
     pub async fn update(context: &Context, user: UpdateUser) -> FieldResult<User> {
         context.authorize_admins_only()?;
+        let mut db = context.db().await?;
 
         let db_user = sqlx::query_as!(
             User,
@@ -63,8 +65,26 @@ impl User {
             ",
             user.user_id.parse::<i32>()?
         )
-            .fetch_one(&mut context.db().await?)
+            .fetch_one(&mut db)
             .await?;
+
+        let admin_count = sqlx::query!(
+            "SELECT COUNT(id) FROM users AS count WHERE is_admin=True"
+        )
+        .fetch_one(&mut db)
+        .await?
+        .count;
+
+        if db_user.is_admin && user.is_admin == Some(false) && admin_count == 1 {
+            let msg = "Cannot remove admin access. Machines must have at least one admin user";
+
+            return Err(FieldError::new(
+                msg,
+                graphql_value!({
+                    "internal_error": msg
+                }),
+            ));
+        };
 
         let next_user = sqlx::query_as!(
             User,
@@ -77,7 +97,7 @@ impl User {
             user.user_id.parse::<i32>()?,
             user.is_admin.unwrap_or(db_user.is_admin)
         )
-            .fetch_one(&mut context.db().await?)
+            .fetch_one(&mut db)
             .await?;
 
         Ok(next_user)
@@ -85,6 +105,7 @@ impl User {
 
     pub async fn delete(context: &Context, user_id: String) -> FieldResult<Option<bool>> {
         println!("{:?}", user_id);
+        let mut db = context.db().await?;
         let user_id = user_id.parse::<i32>()?;
 
         let self_deletion = context.current_user
@@ -97,11 +118,33 @@ impl User {
             context.authorize_admins_only()?;
         };
 
+        let admin_count = sqlx::query!(
+            "SELECT COUNT(id) FROM users AS count WHERE is_admin=True"
+        )
+        .fetch_one(&mut db)
+        .await?
+        .count;
+
+        let admin_deletion= sqlx::query!(
+            "SELECT is_admin FROM users WHERE id=$1",
+            user_id
+        )
+        .fetch_one(&mut db)
+        .await?
+        .is_admin;
+
+        if admin_deletion && admin_count == 1 {
+            return Err(FieldError::new(
+                "Cannot delete only admin user",
+                graphql_value!({ "internal_error": "Cannot delete only admin user" }),
+            ))
+        };
+
         let _ = sqlx::query!(
             "DELETE FROM users WHERE id=$1",
             user_id
         )
-        .fetch_optional(&mut context.db().await?)
+        .fetch_optional(&mut db)
         .await?;
 
         Ok(None)
